+++ Arbiter.App/ViewModels/Entities/EntityManagerViewModel.ContextMenu.cs
@@ -3,6 +3,9 @@ using System.Buffers.Binary;
 using System.Linq;
 using System.Threading.Tasks;
 using Arbiter.App.Extensions;
+using Arbiter.App.Models;
+using Arbiter.Net.Client.Messages;
+using Arbiter.Net.Types;
 using Avalonia;
 using CommunityToolkit.Mvvm.Input;
 
@@ -10,6 +13,25 @@ namespace Arbiter.App.ViewModels.Entities;
 
 public partial class EntityManagerViewModel
 {
+    private bool CanInteract() => SelectedClient is not null && SelectedEntities.Count == 1 &&
+                                  !SelectedEntities[0].Flags.HasFlag(EntityFlags.Item);
+
+    [RelayCommand(CanExecute = nameof(CanInteract))]
+    private void Interact()
+    {
+        if (SelectedClient is null || SelectedEntities.Count == 0)
+        {
+            return;
+        }
+
+        var clientInteract = new ClientInteractMessage
+        {
+            InteractionType = InteractionType.Entity,
+            TargetId =(uint)SelectedEntities[0].Id
+        };
+        SelectedClient.EnqueueMessage(clientInteract);
+    }
+
     private bool CanCopyToClipboard() => SelectedEntities.Count == 1;
     
     [RelayCommand(CanExecute = nameof(CanCopyToClipboard))]